Fix multiple calls to `async.callback` when calling `ahalt` with body
This fixes an issue where the callback would be called multiple times
when calling `ahalt` with a body like so:

    ahalt 404, 'Thing not found'

The problem is apparent when a rack middleware overrides the callback to
perform it's own work:

    def call(env)
      callback = env['async.callback']

      env['async.callback'] = Proc.new do |response|
        puts 'Calling callback!'
        callback.call(response)
      end

      @app.call(env)
    end

This change removes the manual call to `#body` as `#halt` will
automatically do that for you.

It also removes what seems to me like a typo (referencing `#body`
instead of `response.body`) which when called without an argument
would call the super implementation (the default implementation
in Sinatra).